refactor: Split out supertyping from `dtypes`
